[DeadCode] Add SideEffectNodeDetector::detectCallExpr()
