+++ rules/DeadCode/SideEffect/SideEffectNodeDetector.php
@@ -4,16 +4,22 @@ declare(strict_types=1);
 
 namespace Rector\DeadCode\SideEffect;
 
+use PhpParser\Node;
 use PhpParser\Node\Expr;
 use PhpParser\Node\Expr\ArrayDimFetch;
 use PhpParser\Node\Expr\Assign;
 use PhpParser\Node\Expr\BinaryOp\Concat;
 use PhpParser\Node\Expr\FuncCall;
+use PhpParser\Node\Expr\MethodCall;
 use PhpParser\Node\Expr\New_;
+use PhpParser\Node\Expr\NullsafeMethodCall;
 use PhpParser\Node\Expr\PropertyFetch;
+use PhpParser\Node\Expr\StaticCall;
 use PhpParser\Node\Expr\Variable;
+use PhpParser\Node\Name;
 use PhpParser\Node\Scalar\Encapsed;
 use PHPStan\Type\ConstantType;
+use PHPStan\Type\ObjectType;
 use Rector\NodeTypeResolver\NodeTypeResolver;
 
 final class SideEffectNodeDetector
@@ -22,6 +28,15 @@ final class SideEffectNodeDetector
      * @var array<class-string<Expr>>
      */
     private const SIDE_EFFECT_NODE_TYPES = [Encapsed::class, New_::class, Concat::class, PropertyFetch::class];
+
+    /**
+     * @var array<class-string<Expr>>
+     */
+    private const CALL_EXPR_SIDE_EFFECT_NODE_TYPES = [
+        MethodCall::class,
+        NullsafeMethodCall::class,
+        StaticCall::class,
+    ];
 
     public function __construct(
         private NodeTypeResolver $nodeTypeResolver,
@@ -57,6 +72,43 @@ final class SideEffectNodeDetector
         }
 
         return true;
+    }
+
+    public function detectCallExpr(Node $node): bool
+    {
+        if (! $node instanceof Expr) {
+            return false;
+        }
+
+        $exprClass = $node::class;
+
+        if ($node instanceof StaticCall && $this->isClassCallerThrowable($node)) {
+            return false;
+        }
+
+        if (in_array($exprClass, self::CALL_EXPR_SIDE_EFFECT_NODE_TYPES, true)) {
+            return true;
+        }
+
+        if ($node instanceof FuncCall) {
+            return ! $this->pureFunctionDetector->detect($node);
+        }
+
+        return false;
+    }
+
+    private function isClassCallerThrowable(StaticCall $staticCall): bool
+    {
+        $class = $staticCall->class;
+        if (! $class instanceof Name) {
+            return false;
+        }
+
+        $throwableType = new ObjectType('Throwable');
+        $type = new ObjectType($class->toString());
+
+        return $throwableType->isSuperTypeOf($type)
+            ->yes();
     }
 
     private function resolveVariable(Expr $expr): ?Variable
